2nd refactor/cleanup pass
