+++ frontend/package.json
@@ -7,14 +7,18 @@
   "author": "Leigh MacDonald",
   "license": "MIT",
   "private": false,
-  "type": "module",
   "jest": {
     "verbose": true
   },
+  "presets": [
+    [
+      "env",
+      {
+        "modules": false
+      }
+    ]
+  ],
   "scripts": {
-    "compress": "./compress.sh",
-    "copy": "cp -rv src/icons/* ../internal/web/dist/",
-    "copy_win": "xcopy src\\icons\\* ..\\internal\\web\\dist\\ /Y",
     "watch": "cross-env NODE_ENV=development webpack --watch --progress",
     "build": "cross-env NODE_ENV=production webpack",
     "prettier:base": "prettier --parser typescript --single-quote",
@@ -35,17 +39,19 @@
     "@types/chai": "^4.2.21",
     "@types/copy-webpack-plugin": "^8.0.1",
     "@types/es6-promise": "^3.3.0",
-    "@types/jest": "^26.0.24",
+    "@types/jest": "^27.0.1",
     "@types/jquery": "^3.5.6",
     "@types/lodash": "^4.14.171",
     "@types/lodash-es": "^4.17.4",
-    "@types/react": "^17.0.14",
+    "@types/node": "^16.7.10",
+    "@types/react": "^17.0.19",
     "@types/react-dom": "^17.0.9",
     "@types/react-router-dom": "^5.1.8",
+    "@types/steamid": "^1.1.0",
     "@types/webpack": "^5.28.0",
     "@typescript-eslint/eslint-plugin": "^4.28.3",
     "@typescript-eslint/parser": "^4.28.3",
-    "chai": "^4.3.0",
+    "chai": "^4.3.4",
     "copy-webpack-plugin": "^9.0.1",
     "cross-env": "^7.0.2",
     "css-loader": "^6.1.0",
@@ -58,20 +64,24 @@
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "history": "^5.0.0",
+    "html-webpack-plugin": "^5.3.2",
     "ip-cidr": "^3.0.0",
-    "jest": "^27.0.6",
+    "jest": "^27.1.0",
     "js-cookie": "^2.2.1",
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
+    "mini-css-extract-plugin": "^2.2.2",
     "moment": "^2.27.0",
     "node-sass": "^6.0.1",
     "prettier": "^2.3.2",
-    "react": "^17.0.1",
+    "react": "^17.0.2",
     "react-dom": "^17.0.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
+    "react-use-websocket": "^2.7.1",
     "sass": "^1.35.2",
     "sass-loader": "^12.0.0",
+    "steamid": "^1.1.3",
     "style-loader": "^3.1.0",
     "ts-loader": "^9.2.3",
     "ts-node": "^10.1.0",
@@ -80,9 +90,18 @@
     "webpack": "^5.45.1",
     "webpack-cli": "^4.7.2"
   },
-  "dependencies": {
-    "@types/steamid": "^1.1.0",
-    "react-use-websocket": "^2.7.1",
-    "steamid": "^1.1.3"
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all",
+      "ie >= 9"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version",
+      "ie 11"
+    ]
   }
 }
