+++ frontend/src/component/Footer.tsx
@@ -3,7 +3,12 @@ import { Grid, Typography } from '@material-ui/core';
 
 export const Footer = (): JSX.Element => {
     return (
-        <Grid container spacing={3} alignItems="center" justify="space-evenly">
+        <Grid
+            container
+            spacing={3}
+            alignItems="center"
+            justifyContent="space-evenly"
+        >
             <Grid item xs={6}>
                 <Typography align={'center'} variant={'body2'}>
                     <a href="https://github.com/leighmacdonald/gbans">gbans</a>
