+++ frontend/src/component/TopBar.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { fade, makeStyles, Theme } from '@material-ui/core/styles';
+import { alpha, makeStyles, Theme } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import IconButton from '@material-ui/core/IconButton';
@@ -36,6 +36,7 @@ import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 import AccountCircleIcon from '@material-ui/icons/AccountCircle';
 import { GLink } from './GLink';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
+import steamLogo from '../icons/steam_login_sm.png';
 
 const useStyles = makeStyles((theme: Theme) => ({
     grow: {
@@ -53,9 +54,9 @@ const useStyles = makeStyles((theme: Theme) => ({
     search: {
         position: 'relative',
         borderRadius: theme.shape.borderRadius,
-        backgroundColor: fade(theme.palette.common.white, 0.15),
+        backgroundColor: alpha(theme.palette.common.white, 0.15),
         '&:hover': {
-            backgroundColor: fade(theme.palette.common.white, 0.25)
+            backgroundColor: alpha(theme.palette.common.white, 0.25)
         },
         marginRight: theme.spacing(2),
         marginLeft: 0,
@@ -326,9 +327,7 @@ const TopBar = ({ history }: RouteComponentProps): JSX.Element => {
                                 (currentUser?.player.steam_id === '' && (
                                     <Button onClick={handleOnLogin}>
                                         <img
-                                            src={
-                                                '/assets/dist/steam_login_sm.png'
-                                            }
+                                            src={steamLogo}
                                             alt={'Steam Login'}
                                         />
                                     </Button>
