+++ frontend/src/types/types.d.ts
@@ -1,0 +1,11 @@
+declare module '*.png';
+// declare module '*.jpg';
+// declare module '*.jpeg';
+// declare module '*.png';
+// declare module '*.gif';
+// declare module '*.svg';
+// declare module '*.webp';
+// declare module '*.html';
+// declare module '*.webm';
+// declare module '*.zip';
+// declare module '*.css';
