+++ internal/cmd/proxy.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"context"
 	"github.com/leighmacdonald/gbans/internal/proxy"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -13,10 +14,12 @@ var proxyCmd = &cobra.Command{
 	Short: "proxy client",
 	Long:  `gbans game proxy proxy`,
 	Run: func(cmd *cobra.Command, args []string) {
-
+		ctx := context.Background()
 		fa, _ := net.ResolveUDPAddr("udp", "192.168.0.101:10000")
-		ba, _ := net.ResolveUDPAddr("udp", "192.168.0.101:27015")
-		p, err := proxy.New(fa, ba)
+		ba, _ := net.ResolveUDPAddr("udp", "192.168.0.220:27015")
+		p, err := proxy.New(ctx, fa, ba, proxy.Opts{
+			Limit: 75,
+		})
 		if err != nil {
 			log.Errorf("Proxy error: %v", err)
 			return
