+++ internal/cmd/test_data.go
@@ -6,6 +6,7 @@ import (
 	"github.com/leighmacdonald/gbans/internal/config"
 	"github.com/leighmacdonald/gbans/internal/model"
 	"github.com/leighmacdonald/gbans/internal/store"
+	"github.com/leighmacdonald/gbans/pkg/ip2location"
 	"github.com/leighmacdonald/golib"
 	"github.com/leighmacdonald/steamid/v2/steamid"
 	"github.com/leighmacdonald/steamweb"
@@ -22,9 +23,14 @@ type seedData struct {
 	Admins  steamid.Collection `json:"admins"`
 	Players steamid.Collection `json:"players"`
 	Servers []struct {
-		ShortName string `json:"short_name"`
-		Host      string `json:"host"`
-		Password  string `json:"password"`
+		ShortName string    `json:"short_name"`
+		Host      string    `json:"host"`
+		Port      int       `json:"port,omitempty"`
+		Password  string    `json:"password"`
+		Location  []float64 `json:"location"`
+		Enabled   bool      `json:"enabled"`
+		Region    string    `json:"region"`
+		CC        string    `json:"cc"`
 	} `json:"servers"`
 	Settings struct {
 		Rcon string `json:"rcon"`
@@ -56,7 +62,10 @@ var testDataCmd = &cobra.Command{
 
 		for _, adminSid := range seed.Admins {
 			var p model.Person
-			_ = db.GetOrCreatePersonBySteamID(ctx, adminSid, &p)
+			if errGP := db.GetOrCreatePersonBySteamID(ctx, adminSid, &p); errGP != nil {
+				log.Fatalf("Failed to get person: %v", errGP)
+			}
+
 			sum1, err := steamweb.PlayerSummaries(steamid.Collection{p.SteamID})
 			if err != nil {
 				log.Errorf("Failed to get player summary: %v", err)
@@ -84,6 +93,20 @@ var testDataCmd = &cobra.Command{
 				return
 			}
 		}
+		fl, efl := steamweb.GetFriendList(76561197961279983)
+		if efl != nil {
+			log.Errorf("Failed to get friendlist")
+		}
+		for _, f := range fl {
+			p := model.NewPerson(f.Steamid)
+			if errF := db.GetOrCreatePersonBySteamID(ctx, f.Steamid, &p); errF != nil {
+				log.Errorf("Failed to create person: %v", errF)
+			}
+			b := model.NewBan(p.SteamID, seed.Players[0], time.Hour*500)
+			if errB := db.SaveBan(ctx, &b); errB != nil {
+				log.Errorf("Failed to make ban: %v", errB)
+			}
+		}
 		for _, server := range seed.Servers {
 			pw := golib.RandomString(20)
 			if server.Password != "" {
@@ -93,14 +116,25 @@ var testDataCmd = &cobra.Command{
 			if testRconPass != "" {
 				rconPass = testRconPass
 			}
+			port := 27015
+			if server.Port > 0 {
+				port = server.Port
+			}
 			s := model.Server{
-				ServerName:     server.ShortName,
-				Address:        server.Host,
-				Port:           27015,
-				RCON:           rconPass,
-				ReservedSlots:  8,
-				Token:          "0123456789012345678901234567890123456789",
-				Password:       pw,
+				ServerName:    server.ShortName,
+				Token:         golib.RandomString(40),
+				Address:       server.Host,
+				Port:          port,
+				RCON:          rconPass,
+				ReservedSlots: 8,
+				Password:      pw,
+				IsEnabled:     server.Enabled,
+				Region:        server.Region,
+				CC:            server.CC,
+				Location: ip2location.LatLong{
+					Latitude:  server.Location[0],
+					Longitude: server.Location[1],
+				},
 				TokenCreatedOn: config.Now(),
 				CreatedOn:      config.Now(),
 				UpdatedOn:      config.Now(),
