+++ internal/store/db_stats.go
@@ -1,0 +1,34 @@
+package store
+
+import (
+	"context"
+	"github.com/leighmacdonald/gbans/internal/model"
+	log "github.com/sirupsen/logrus"
+)
+
+func (db *pgStore) GetStats(ctx context.Context, stats *model.Stats) error {
+	const q = `
+	SELECT 
+		(SELECT COUNT(ban_id) FROM ban) as bans_total,
+		(SELECT COUNT(ban_id) FROM ban WHERE created_on >= (now() - INTERVAL '1 DAY')) as bans_day,
+	    (SELECT COUNT(ban_id) FROM ban WHERE created_on >= (now() - INTERVAL '1 DAY')) as bans_week,
+		(SELECT COUNT(ban_id) FROM ban WHERE created_on >= (now() - INTERVAL '1 MONTH')) as bans_month, 
+	    (SELECT COUNT(ban_id) FROM ban WHERE created_on >= (now() - INTERVAL '3 MONTH')) as bans_3month,
+	    (SELECT COUNT(ban_id) FROM ban WHERE created_on >= (now() - INTERVAL '6 MONTH')) as bans_6month,
+	    (SELECT COUNT(ban_id) FROM ban WHERE created_on >= (now() - INTERVAL '1 YEAR')) as bans_year,
+		(SELECT COUNT(net_id) FROM ban_net) as bans_cidr, 
+		(SELECT COUNT(appeal_id) FROM ban_appeal WHERE appeal_state = 0) as appeals_open,
+		(SELECT COUNT(appeal_id) FROM ban_appeal WHERE appeal_state = 1 OR appeal_state = 2) as appeals_closed,
+		(SELECT COUNT(word_id) FROM filtered_word) as filtered_words,
+		(SELECT COUNT(server_id) FROM server) as servers_total`
+	if err := db.c.QueryRow(ctx, q).
+		Scan(&stats.BansTotal, &stats.BansDay, &stats.BansWeek, &stats.BansMonth,
+			&stats.Bans3Month, &stats.Bans6Month, &stats.BansYear, &stats.BansCIDRTotal,
+			&stats.AppealsOpen, &stats.AppealsClosed, &stats.FilteredWords, &stats.ServersTotal,
+		); err != nil {
+		log.Errorf("Failed to fetch stats: %v", err)
+		return dbErr(err)
+	}
+	return nil
+
+}
