+++ internal/store/store.go
@@ -13,10 +13,17 @@ import (
 
 type ServerStore interface {
 	GetServer(ctx context.Context, serverID int64, s *model.Server) error
-	GetServers(ctx context.Context) ([]model.Server, error)
+	GetServers(ctx context.Context, includeDisabled bool) ([]model.Server, error)
 	GetServerByName(ctx context.Context, serverName string, s *model.Server) error
 	SaveServer(ctx context.Context, server *model.Server) error
 	DropServer(ctx context.Context, serverID int64) error
+}
+
+type DemoStore interface {
+	GetDemo(ctx context.Context, demoId int64, d *model.DemoFile) error
+	GetDemos(ctx context.Context) ([]model.DemoFile, error)
+	SaveDemo(ctx context.Context, d *model.DemoFile) error
+	DropDemo(ctx context.Context, d *model.DemoFile) error
 }
 
 type BanStore interface {
@@ -27,10 +34,10 @@ type BanStore interface {
 	SaveBan(ctx context.Context, ban *model.Ban) error
 	GetBanNet(ctx context.Context, ip net.IP) ([]model.BanNet, error)
 	SaveBanNet(ctx context.Context, banNet *model.BanNet) error
-	DropNetBan(ctx context.Context, ban model.BanNet) error
+	DropNetBan(ctx context.Context, ban *model.BanNet) error
 	DropBan(ctx context.Context, ban *model.Ban) error
-	GetExpiredBans(ctx context.Context) ([]*model.Ban, error)
-	GetBans(ctx context.Context, o *QueryFilter) ([]*model.BannedPerson, error)
+	GetExpiredBans(ctx context.Context) ([]model.Ban, error)
+	GetBans(ctx context.Context, o *QueryFilter) ([]model.BannedPerson, error)
 	GetBansOlderThan(ctx context.Context, o *QueryFilter, t time.Time) ([]model.Ban, error)
 	GetExpiredNetBans(ctx context.Context) ([]model.BanNet, error)
 	Import(ctx context.Context, root string) error
@@ -40,11 +47,12 @@ type PersonStore interface {
 	DropPerson(ctx context.Context, steamID steamid.SID64) error
 	SavePerson(ctx context.Context, person *model.Person) error
 	GetPersonBySteamID(ctx context.Context, sid steamid.SID64, p *model.Person) error
-	GetPeople(ctx context.Context, qf *QueryFilter) ([]*model.Person, error)
+	GetPeople(ctx context.Context, qf *QueryFilter) ([]model.Person, error)
 	GetOrCreatePersonBySteamID(ctx context.Context, sid steamid.SID64, p *model.Person) error
 	GetPersonByDiscordID(ctx context.Context, did string, p *model.Person) error
 	AddPersonIP(ctx context.Context, p *model.Person, ip string) error
 	GetIPHistory(ctx context.Context, sid64 steamid.SID64) ([]model.PersonIPRecord, error)
+	GetExpiredProfiles(ctx context.Context, limit int) ([]model.Person, error)
 }
 
 type FilterStore interface {
@@ -76,6 +84,7 @@ type NetworkStore interface {
 // Store defines our composite store interface encapsulating all store interfaces
 type Store interface {
 	BanStore
+	DemoStore
 	FilterStore
 	MigrationStore
 	NetworkStore
